+++ mchap/application/baseclass.py
@@ -69,13 +69,14 @@ class program(object):
         infofields = [
             "AN",
             "AC",
+            "REFMASKED",
             "NS",
             "DP",
             "RCOUNT",
             "END",
             "NVAR",
             "SNVPOS",
-        ] + [f for f in ["AFP"] if f in self.report_fields]
+        ] + [f for f in ["AFPRIOR", "AFP"] if f in self.report_fields]
         return infofields
 
     def format_fields(self):
@@ -113,7 +114,9 @@ class program(object):
         ]
         contigs = self.header_contigs()
         filters = [
-            vcf.filters.SamplePassFilter(),
+            vcf.filters.PASS,
+            vcf.filters.NOA,
+            vcf.filters.AF0,
         ]
         info_fields = [HEADER_INFO_FIELDS[field] for field in self.info_fields()]
         format_fields = [HEADER_FORMAT_FIELDS[field] for field in self.format_fields()]
@@ -135,7 +138,7 @@ class program(object):
             sample_inbreeding=self.sample_inbreeding,
             infofields=infofields,
             formatfields=formatfields,
-            columndata=dict(),
+            columndata=dict(FILTER=list()),
             infodata=dict(),
             sampledata=dict(),
         )
@@ -276,7 +279,7 @@ class program(object):
         Returns
         -------
         data : LocusAssemblyData
-            With columndata fields: "REF" "ALTS" and infodata fields:
+            With infodata fields:
             "END", "NVAR", "SNVPOS", "AC", "AN", "NS", "DP", "RCOUNT"
             and "AFP" if specified.
         """
@@ -286,9 +289,9 @@ class program(object):
         data.infodata["SNVPOS"] = (
             np.subtract(data.locus.positions, data.locus.start) + 1
         )
-        # sequences
-        data.columndata["REF"] = data.locus.sequence
-        data.columndata["ALTS"] = data.locus.alts
+        # if no filters applied then locus passed
+        if len(data.columndata["FILTER"]) == 0:
+            data.columndata["FILTER"] = vcf.filters.PASS.id
         # alt allele counts
         allele_counts = np.zeros(len(data.columndata["ALTS"]) + 1, int)
         for array in data.sampledata["alleles"].values():