+++ mchap/io/__init__.py
@@ -1,6 +1,6 @@
 #!/usr/bin/env python3
 
-from .loci import SNP, Locus, read_bed4
+from .loci import SNP, Locus, LocusPrior, read_bed4
 from .util import qual_of_char, prob_of_qual, qual_of_prob
 from .bam import (
     extract_sample_ids,
@@ -12,6 +12,7 @@ from .bam import (
 __all__ = [
     "SNP",
     "Locus",
+    "LocusPrior",
     "read_bed4",
     "qual_of_char",
     "prob_of_qual",