+++ mchap/tests/test_application_assemble.py
@@ -235,6 +235,8 @@ def test_Program__header():
 
     filters_expect = [
         '##FILTER=<ID=PASS,Description="All filters passed">',
+        '##FILTER=<ID=NOA,Description="No observed alleles at locus">',
+        '##FILTER=<ID=AF0,Description="All alleles have prior allele frequency of zero">',
     ]
     filters_actual = [line for line in header if line.startswith("##FILTER")]
     assert filters_actual == filters_expect